+++ resources/release-notes.md
@@ -1,3 +1,4 @@
 ## Software
 
+- (All) Make `mag.Science/select` more flexible for missions other than IMAP
 - (IMAP) Adjust health check limits for 1.8 V and 3.3 V currents based on in-flight data